move common fixtures to dedicated directory and update cleanup function to use sweeper module
