+++ tests/integration/fixtures/__init__.py
@@ -1,0 +1,5 @@
+from .containerclusters import *
+from .labels import *
+from .rulesets import *
+from .services import *
+from .workloads import *
